@@ -28,15 +28,12 @@ package_dir =
     =py
 packages = find:
 include_package_data = True
-python_requires = >=3.9
+python_requires = >=3.10
 # setup_requires = setuptools_scm
 install_requires =
     requests
     pyyaml
-    numpy<2.0
-    astropy>=5.0
-    healpy
-    matplotlib
+    astropy<7
 scripts =
     bin/annotate_fits
     bin/desi_api_file
@@ -56,11 +53,16 @@ scripts =
 #     desi_update_version = desiutil.setup:main
 
 [options.extras_require]
+all =
+    healpy
+    matplotlib
+    scipy
+    speclite
 test =
+    desiutil[all]
     pytest
-    scipy
 coverage =
-    scipy
+    desiutil[all]
     pytest-cov
     coveralls
 doc =
@@ -115,9 +117,13 @@ exclude_lines =
 # These are normally ignored by default:
 # ignore = E121, E123, E126, E133, E226, E241, E242, E704, W503, W504
 #
-# In addition to the default set we add:
+# These are the explicitly ignored styles:
+# - E226: missing whitespace around arithmetic operator
+# - E241: multiple spaces after ','
+# - E741: ambiguous variable name 'l', see desutil.dust._Hemisphere.ebv().
+# - W504: line break after binary operator
+#
+# And finally this one is not ignored, but modified by max-line-length.
 # - E501: line too long (82 > 79 characters)
-# - E731: do not assign a lambda expression, use a def
-# - E741: do not use variables named 'l', 'O', or 'I' -- because, for example,
-#   'l' might refer to Galactic longitude.
-ignore = E121, E123, E126, E133, E226, E241, E242, E501, E704, E731, E741, W503, W504
+max-line-length = 160
+ignore = E226, E241, E741, W504
