@@ -52,7 +52,7 @@ def freeze_iers(name='iers_frozen.ecsv', ignore_warnings=True):
     :meth:`astropy.utils.iers.IERS._check_interpolate_indices` to prevent
     any IERSRangeError being raised.
 
-    See http://docs.astropy.org/en/stable/utils/iers.html for details.
+    See https://docs.astropy.org/en/stable/utils/iers.html for details.
 
     This function returns immediately after the first time it is called,
     so it it safe to insert anywhere that consistent IERS models are
@@ -105,7 +105,8 @@ def freeze_iers(name='iers_frozen.ecsv', ignore_warnings=True):
     # to prevent any IERSRangeError being raised.
     class IERS_Frozen(astropy.utils.iers.IERS_B):
         def _check_interpolate_indices(self, indices_orig, indices_clipped,
-                                       max_input_mjd): pass
+                                       max_input_mjd):
+            pass
 
     # Create and register an instance of this class from the table.
     iers = IERS_Frozen(table)
@@ -121,9 +122,15 @@ def freeze_iers(name='iers_frozen.ecsv', ignore_warnings=True):
     else:
         astropy.utils.iers.conf.iers_degraded_accuracy = 'warn'
     # Sanity check.
-    auto_class = astropy.utils.iers.IERS_Auto.open()
-    if auto_class is not iers:
-        raise RuntimeError('Frozen IERS is not installed as the default ({0} v. {1}).'.format(auto_class.__class__, iers.__class__))
+    # In Astropy 7 this appears to be broken due to the iers_frozen.ecsv being out of date.
+    # The *format* of that file no longer matches what is expected by astropy.util.iers.
+    try:
+        auto_class = astropy.utils.iers.IERS_Auto.open()
+        if auto_class is not iers:
+            raise RuntimeError('Frozen IERS is not installed as the default ({0} v. {1}).'.format(auto_class.__class__, iers.__class__))
+    except KeyError:
+        # Temporary Astropy 7/IERS workaround.
+        warnings.warn("Temporarily skipping IERS integrity check.", UserWarning)
 
     if ignore_warnings:
         try:
