@@ -22,13 +22,16 @@ class TestNames(unittest.TestCase):
     def test_radec_to_desiname(self):
         """Test computation of desiname.
         """
-        ras = [6.2457354547234, 23.914121939862518, 36.23454570972834,
+        ras = [0.003456789, 6.2457354547234, 23.914121939862518, 36.23454570972834, 40.669625, 40.669625,
                235.25235223446, 99.9999999999999]
-        decs = [29.974787585945496, -42.945872347904356, -0.9968423456,
+        decs = [7.434432123, 29.974787585945496, -42.945872347904356, -0.9968423456, -0.013277777, 0.013277777,
                 8.45677345352345, 89.234958294953]
-        correct_names = np.array(['DESI J006.2457+29.9747',
+        correct_names = np.array(['DESI J000.0034+07.4344',
+                                  'DESI J006.2457+29.9747',
                                   'DESI J023.9141-42.9458',
                                   'DESI J036.2345-00.9968',
+                                  'DESI J040.6696-00.0132',
+                                  'DESI J040.6696+00.0132',
                                   'DESI J235.2523+08.4567',
                                   'DESI J099.9999+89.2349'])
         # Test scalar conversion
@@ -38,12 +41,12 @@ class TestNames(unittest.TestCase):
 
         # Test list conversion
         outnames = radec_to_desiname(ras, decs)
-        self.assertTrue(np.alltrue(outnames == correct_names))
+        self.assertTrue((outnames == correct_names).all())
 
         # Test array conversion
         outnames = radec_to_desiname(np.array(ras),
                                      np.array(decs))
-        self.assertTrue(np.alltrue(outnames == correct_names))
+        self.assertTrue((outnames == correct_names).all())
 
     def test_radec_to_desiname_bad_values(self):
         """Test exceptions when running radec_to_desiname with bad values.
